New suggestion engine based on reducers
