+++ test/elixir_sense/suggestions_test.exs
@@ -56,7 +56,6 @@ defmodule ElixirSense.SuggestionsTest do
     list = ElixirSense.suggestions(buffer, 2, 7)
 
     assert list == [
-             %{type: :hint, value: "is_b"},
              %{
                args: "term",
                arity: 1,
@@ -103,7 +102,6 @@ defmodule ElixirSense.SuggestionsTest do
     list = ElixirSense.suggestions(buffer, 3, 14)
 
     assert list == [
-             %{type: :hint, value: "MyList.flatten"},
              %{
                args: "list",
                arity: 1,
@@ -140,7 +138,6 @@ defmodule ElixirSense.SuggestionsTest do
     list = ElixirSense.suggestions(buffer, 3, 12)
 
     assert list == [
-             %{type: :hint, value: "Macros.some"},
              %{
                args: "var",
                arity: 1,
@@ -164,7 +161,6 @@ defmodule ElixirSense.SuggestionsTest do
     list = ElixirSense.suggestions(buffer, 2, 34)
 
     assert list == [
-             %{type: :hint, value: "ElixirSenseExample.ModuleWithDoc"},
              %{
                name: "ModuleWithDocFalse",
                subtype: nil,
@@ -897,7 +893,7 @@ defmodule ElixirSense.SuggestionsTest do
       ElixirSense.suggestions(buffer, 2, 37)
       |> Enum.filter(fn s -> s.type in [:variable, :hint] end)
 
-    assert list == [%{type: :hint, value: "my_var"}, %{name: "my_var", type: :variable}]
+    assert list == [%{name: "my_var", type: :variable}]
   end
 
   test "variable shadowing function" do
@@ -912,7 +908,6 @@ defmodule ElixirSense.SuggestionsTest do
     """
 
     assert [
-             %{type: :hint, value: "my_fun"},
              %{name: "my_fun", type: :variable},
              %{name: "my_fun", type: :function}
            ] = ElixirSense.suggestions(buffer, 5, 9)
@@ -993,7 +988,6 @@ defmodule ElixirSense.SuggestionsTest do
     """
 
     assert [
-             %{type: :hint, value: "test_fun_p"},
              %{
                arity: 0,
                name: "test_fun_priv",
@@ -1009,7 +1003,6 @@ defmodule ElixirSense.SuggestionsTest do
            ] = ElixirSense.suggestions(buffer, 5, 7)
 
     assert [
-             %{type: :hint, value: "test_fun_priv"},
              %{
                arity: 0,
                name: "test_fun_priv",
@@ -1019,7 +1012,6 @@ defmodule ElixirSense.SuggestionsTest do
            ] = ElixirSense.suggestions(buffer, 6, 21)
 
     assert [
-             %{type: :hint, value: "is_boo"},
              %{
                arity: 1,
                name: "is_boolean",
@@ -1035,7 +1027,6 @@ defmodule ElixirSense.SuggestionsTest do
            ] = ElixirSense.suggestions(buffer, 7, 10)
 
     assert [
-             %{type: :hint, value: "delegate_"},
              %{
                arity: 0,
                name: "delegate_defined",
@@ -1051,7 +1042,6 @@ defmodule ElixirSense.SuggestionsTest do
            ] = ElixirSense.suggestions(buffer, 8, 8)
 
     assert [
-             %{type: :hint, value: "my_guard_p"},
              %{
                args: "value",
                arity: 1,
@@ -1088,7 +1078,6 @@ defmodule ElixirSense.SuggestionsTest do
     """
 
     assert [
-             %{type: :hint, value: "ElixirSenseExample.ModuleO.test_fun_pub"},
              %{
                arity: 1,
                name: "test_fun_pub",
@@ -1114,7 +1103,6 @@ defmodule ElixirSense.SuggestionsTest do
     """
 
     assert [
-             %{type: :hint, value: "ModuleO.test_fun_pub"},
              %{
                arity: 1,
                name: "test_fun_pub",
@@ -1142,7 +1130,6 @@ defmodule ElixirSense.SuggestionsTest do
     """
 
     assert [
-             %{type: :hint, value: "test_fun_pub"},
              %{
                arity: 1,
                name: "test_fun_pub",
@@ -1156,7 +1143,7 @@ defmodule ElixirSense.SuggestionsTest do
            ] == ElixirSense.suggestions(buffer, 10, 7)
 
     # builtin functions not called locally
-    assert [%{type: :hint, value: "__i"}] == ElixirSense.suggestions(buffer, 11, 8)
+    assert [] == ElixirSense.suggestions(buffer, 11, 8)
   end
 
   test "functions and module suggestions with __MODULE__" do
@@ -1178,7 +1165,6 @@ defmodule ElixirSense.SuggestionsTest do
     """
 
     assert [
-             %{type: :hint, value: "__MODULE__.SmodO"},
              %{
                name: "SmodO",
                type: :module
@@ -1186,7 +1172,6 @@ defmodule ElixirSense.SuggestionsTest do
            ] = ElixirSense.suggestions(buffer, 9, 18)
 
     assert [
-             %{type: :hint, value: "__MODULE__.SmodO.test_fun_pub"},
              %{
                arity: 1,
                name: "test_fun_pub",
@@ -1196,10 +1181,9 @@ defmodule ElixirSense.SuggestionsTest do
            ] = ElixirSense.suggestions(buffer, 10, 24)
 
     # no private on external call
-    assert [%{type: :hint, value: "__MODULE__.te"}] = ElixirSense.suggestions(buffer, 11, 18)
+    assert [] = ElixirSense.suggestions(buffer, 11, 18)
 
     assert [
-             %{type: :hint, value: "__MODULE__.__info__"},
              %{
                arity: 1,
                name: "__info__",
@@ -1218,9 +1202,7 @@ defmodule ElixirSense.SuggestionsTest do
 
     list = ElixirSense.suggestions(buffer, 2, 5)
 
-    assert Enum.at(list, 0) == %{type: :hint, value: "Elixir"}
-
-    assert Enum.at(list, 1) == %{
+    assert Enum.at(list, 0) == %{
              type: :module,
              name: "Elixir",
              subtype: nil,
@@ -1238,7 +1220,7 @@ defmodule ElixirSense.SuggestionsTest do
     """
 
     list = ElixirSense.suggestions(buffer, 3, 9)
-    assert Enum.at(list, 1).name == "is_odd"
+    assert Enum.at(list, 0).name == "is_odd"
   end
 
   test "suggestion for struct fields" do
@@ -1251,10 +1233,9 @@ defmodule ElixirSense.SuggestionsTest do
 
     list =
       ElixirSense.suggestions(buffer, 2, 14)
-      |> Enum.filter(&(&1.type in [:field, :hint]))
+      |> Enum.filter(&(&1.type in [:field]))
 
     assert list == [
-             %{type: :hint, value: ""},
              %{
                name: "device",
                origin: "IO.Stream",
@@ -1287,10 +1268,9 @@ defmodule ElixirSense.SuggestionsTest do
 
     list =
       ElixirSense.suggestions(buffer, 3, 18)
-      |> Enum.filter(&(&1.type in [:field, :hint]))
+      |> Enum.filter(&(&1.type in [:field]))
 
     assert list == [
-             %{type: :hint, value: ""},
              %{
                name: "__exception__",
                origin: "ArgumentError",
@@ -1328,7 +1308,6 @@ defmodule ElixirSense.SuggestionsTest do
       |> Enum.filter(&(&1.type in [:field, :hint]))
 
     assert list == [
-             %{type: :hint, value: ""},
              %{
                name: "device",
                origin: "IO.Stream",
@@ -1373,7 +1352,6 @@ defmodule ElixirSense.SuggestionsTest do
       |> Enum.filter(&(&1.type in [:field, :hint]))
 
     assert list == [
-             %{type: :hint, value: ""},
              %{
                name: "__struct__",
                origin: nil,
@@ -1388,7 +1366,6 @@ defmodule ElixirSense.SuggestionsTest do
       |> Enum.filter(&(&1.type in [:field, :hint]))
 
     assert list == [
-             %{type: :hint, value: ""},
              %{
                name: "__struct__",
                origin: nil,
@@ -1412,7 +1389,6 @@ defmodule ElixirSense.SuggestionsTest do
       |> Enum.filter(&(&1.type in [:field, :hint]))
 
     assert list == [
-             %{type: :hint, value: ""},
              %{
                name: "device",
                origin: "IO.Stream",
@@ -1464,7 +1440,6 @@ defmodule ElixirSense.SuggestionsTest do
       |> Enum.filter(&(&1.type in [:field, :hint]))
 
     assert list == [
-             %{type: :hint, value: ""},
              %{
                name: "field_1",
                origin: "MyServer",
@@ -1491,7 +1466,6 @@ defmodule ElixirSense.SuggestionsTest do
     list = ElixirSense.suggestions(buffer, 9, 28)
 
     assert list == [
-             %{type: :hint, value: ""},
              %{
                name: "field_1",
                origin: "MyServer",
@@ -1529,7 +1503,6 @@ defmodule ElixirSense.SuggestionsTest do
       |> Enum.filter(&(&1.type in [:field, :hint]))
 
     assert list == [
-             %{type: :hint, value: ""},
              %{
                name: "field_1",
                origin: ":my_server",
@@ -1556,7 +1529,6 @@ defmodule ElixirSense.SuggestionsTest do
     list = ElixirSense.suggestions(buffer, 9, 30)
 
     assert list == [
-             %{type: :hint, value: ""},
              %{
                name: "field_1",
                origin: ":my_server",
@@ -1594,7 +1566,6 @@ defmodule ElixirSense.SuggestionsTest do
     list = ElixirSense.suggestions(buffer, 10, 7)
 
     assert list == [
-             %{type: :hint, value: ""},
              %{
                name: "field_1",
                origin: "MyServer",
@@ -1629,7 +1600,6 @@ defmodule ElixirSense.SuggestionsTest do
     list = ElixirSense.suggestions(buffer, 8, 31)
 
     assert list == [
-             %{type: :hint, value: ""},
              %{
                name: "field_1",
                origin: "MyServer",
@@ -1667,7 +1637,6 @@ defmodule ElixirSense.SuggestionsTest do
       |> Enum.filter(&(&1.type in [:field, :hint]))
 
     assert list == [
-             %{type: :hint, value: "var_1.field_"},
              %{
                name: "field_1",
                origin: "MyServer",
@@ -1700,7 +1669,6 @@ defmodule ElixirSense.SuggestionsTest do
       |> Enum.filter(&(&1.type in [:field, :hint]))
 
     assert list == [
-             %{type: :hint, value: "var_1.key_"},
              %{name: "key_1", origin: nil, type: :field, call?: true, subtype: :map_key},
              %{name: "key_2", origin: nil, type: :field, call?: true, subtype: :map_key}
            ]
@@ -1721,7 +1689,6 @@ defmodule ElixirSense.SuggestionsTest do
       |> Enum.filter(&(&1.type in [:field, :hint]))
 
     assert list == [
-             %{type: :hint, value: "@var_1.key_"},
              %{name: "key_1", origin: nil, type: :field, call?: true, subtype: :map_key},
              %{name: "key_2", origin: nil, type: :field, call?: true, subtype: :map_key}
            ]
@@ -1742,7 +1709,6 @@ defmodule ElixirSense.SuggestionsTest do
       |> Enum.filter(&(&1.type in [:function, :hint]))
 
     assert [
-             %{type: :hint, value: "var_1.to_string"},
              %{name: "to_string", origin: "Atom", type: :function}
            ] = list
   end
@@ -1764,15 +1730,14 @@ defmodule ElixirSense.SuggestionsTest do
     list = ElixirSense.suggestions(buffer, 8, 17)
 
     assert list == [
-             %{type: :hint, value: "some_"},
-             %{name: "some_arg", type: :variable},
              %{
                origin: "MyServer",
                type: :field,
                name: "some_field",
                call?: false,
                subtype: :struct_field
-             }
+             },
+             %{name: "some_arg", type: :variable}
            ]
   end
 
@@ -1793,7 +1758,6 @@ defmodule ElixirSense.SuggestionsTest do
     list = ElixirSense.suggestions(buffer, 8, 28)
 
     assert list == [
-             %{type: :hint, value: "field_1"},
              %{
                call?: false,
                name: "field_1",
@@ -1821,7 +1785,6 @@ defmodule ElixirSense.SuggestionsTest do
     list = ElixirSense.suggestions(buffer, 8, 20)
 
     assert list == [
-             %{type: :hint, value: "field_1"},
              %{
                call?: false,
                name: "field_1",
@@ -1849,7 +1812,6 @@ defmodule ElixirSense.SuggestionsTest do
     list = ElixirSense.suggestions(buffer, 9, 14)
 
     assert list == [
-             %{type: :hint, value: "field_1"},
              %{
                call?: false,
                name: "field_1",
@@ -1872,7 +1834,6 @@ defmodule ElixirSense.SuggestionsTest do
     list = ElixirSense.suggestions(buffer, 3, 20)
 
     assert list == [
-             %{type: :hint, value: "field_1"},
              %{call?: false, name: "field_1", origin: nil, subtype: :struct_field, type: :field}
            ]
   end
@@ -1888,7 +1849,6 @@ defmodule ElixirSense.SuggestionsTest do
     list = ElixirSense.suggestions(buffer, 3, 9)
 
     assert list == [
-             %{type: :hint, value: "hour"},
              %{call?: false, name: "hour", origin: "Time", subtype: :struct_field, type: :field}
            ]
   end
@@ -1905,7 +1865,6 @@ defmodule ElixirSense.SuggestionsTest do
     list = ElixirSense.suggestions(buffer, 3, 20)
 
     assert list == [
-             %{type: :hint, value: "field_1"},
              %{call?: false, name: "field_1", origin: nil, subtype: :map_key, type: :field}
            ]
   end
@@ -1930,7 +1889,6 @@ defmodule ElixirSense.SuggestionsTest do
     list = ElixirSense.suggestions(buffer, 11, 18)
 
     assert list == [
-             %{type: :hint, value: "other_"},
              %{name: "other_arg", type: :variable},
              %{
                name: "other_func",
@@ -1983,8 +1941,7 @@ defmodule ElixirSense.SuggestionsTest do
 
     list = ElixirSense.suggestions(buffer, 2, 22)
 
-    assert [%{type: :hint, value: "__MODULE__.Reducers"}, %{name: "Reducers", type: :module} | _] =
-             list
+    assert [%{name: "Reducers", type: :module} | _] = list
   end
 
   test "suggest modules to alias v1.2 syntax" do
@@ -1996,7 +1953,7 @@ defmodule ElixirSense.SuggestionsTest do
 
     list = ElixirSense.suggestions(buffer, 2, 19)
 
-    assert [%{type: :hint, value: "Reducers"}, %{name: "Reducers", type: :module}] = list
+    assert [%{name: "Reducers", type: :module}] = list
   end
 
   test "suggest modules to alias v1.2 syntax with __MODULE__" do
@@ -2008,7 +1965,7 @@ defmodule ElixirSense.SuggestionsTest do
 
     list = ElixirSense.suggestions(buffer, 2, 23)
 
-    assert [%{type: :hint, value: "Reducers"}, %{name: "Reducers", type: :module}] = list
+    assert [%{name: "Reducers", type: :module}] = list
   end
 
   describe "suggestion for param options" do
@@ -2641,7 +2598,6 @@ defmodule ElixirSense.SuggestionsTest do
     """
 
     assert [
-             %{type: :hint, value: "SameModule.test_fun"},
              %{origin: "ElixirSenseExample.SameModule"}
            ] = ElixirSense.suggestions(buffer, 4, 17)
 
@@ -2657,7 +2613,6 @@ defmodule ElixirSense.SuggestionsTest do
     """
 
     assert [
-             %{type: :hint, value: "SameModule.test_fun"},
              %{origin: "ElixirSenseExample.SameModule"}
            ] = ElixirSense.suggestions(buffer, 5, 17)
 
@@ -2670,7 +2625,6 @@ defmodule ElixirSense.SuggestionsTest do
     """
 
     assert [
-             %{type: :hint, value: "SameModule.some_test_macro"},
              %{origin: "ElixirSenseExample.SameModule"}
            ] = ElixirSense.suggestions(buffer, 4, 15)
   end
